Add support for unmangling cited links
Hook into the composer window and scan the cited text on load for any
mangled links. Works both for plain-text and HTML mode.

This requires some jumping through hoops because the load event comes
too early, and the content is not yet available. Setting a miminal
timeout works around this. Would be nice to know if there is a cleaner
way...

Signed-off-by: Jan Kiszka <jan.kiszka@web.de>